+++ port/src/Permutation.cpp
@@ -10,8 +10,8 @@ void Permutation::setIdentity(std::size_t size) {
     mOldToNew.resize(size);
     mNewToOld.resize(size);
     for (std::size_t i = 0; i < size; ++i) {
-        mOldToNew[i] = i;
-        mNewToOld[i] = i;
+        mOldToNew[i] = static_cast<std::int32_t>(i);
+        mNewToOld[i] = static_cast<std::int32_t>(i);
     }
 }
 
@@ -19,7 +19,7 @@ void Permutation::rebuildInverse() {
     const std::size_t n = mOldToNew.size();
     mNewToOld.assign(n, 0);
     for (std::size_t i = 0; i < n; ++i)
-        mNewToOld[mOldToNew[i]] = i;
+        mNewToOld[static_cast<std::size_t>(mOldToNew[i])] = static_cast<std::int32_t>(i);
 }
 
 bool Permutation::validate() const {
@@ -31,11 +31,11 @@ bool Permutation::validate() const {
     // maps are consistent inverses.
     std::vector<bool> seen(n, false);
     for (std::size_t i = 0; i < n; ++i) {
-        const std::size_t j = mOldToNew[i];
+        const std::size_t j = static_cast<std::size_t>(mOldToNew[i]);
         if (j >= n || seen[j])
             return false;
         seen[j] = true;
-        if (mNewToOld[j] != i)
+        if (static_cast<std::size_t>(mNewToOld[j]) != i)
             return false;
     }
     return true;
